+++ pkgs/google_cloud_storage/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.6.4-wip
 
-* Send `gccl/<version>` in `x-goog-api-client` request header for API usage attribution.
-* Require `google_cloud_rpc: ^0.5.5`.
+* Send `gccl/<version>` in `x-goog-api-client` request header for API usage
+  attribution.
+* Require `google_cloud_rpc: ^0.6.0`.
 * Add an `ifMetagenerationNotMatch` parameter to `Storage.patchBucket`,
   `Storage.uploadObject`, and `Storage.uploadObjectFromString`. If the
   precondition is not satisfied, a `NotModifiedException` is thrown.